Always set false noncontig in internal win_allocate_shared.
To support noncontig hint, we need to modify the internal shared window
creation in win_allocate: (1) allowing arbitrary order of segments in
offset calculation; (2) using win_create_dynamic instead of win_create,
to attach multiple segments on ghosts. Moreover, we need to optimize
win_create_dynamic in MPI implementation similar as that for win_create.

The above work is planed as a future work. For now, the Casper library
just always sets noncontigu to false in the internal
win_allocate_shared.