Added mpl timer option for internal timestamps.
Use PMPI_Wtime by default, user can enable the MPL timer by setting
--with-timer=mpl and --with-mpl=[path] at configure time.

We note that the default MPI timer might use the slow but portable
gettimeofday call. This can result in significant overhead when the
dynamic adaptation in Casper is enabled. Thus it would be good to use
a faster external timer. MPL is included in the MPICH source package
and provides various types of timers.

TODO: the MPL installation requires some manual change to expose timer
APIs. We might need to add it with the embedded mode.